Add tentative drill patterns for Tayda enclosure drilling
Fix up the layout so that the same drill pattern with three holes for
pots/switches works for both the Baxandall and Boost, and move some of
the other controls around to make for whole numbers in mm.

Do the drill patterns work? They match the PCB layout, but let's see
about the vertical sides...  Tayda drilling is running behind, so I
don't think I'll see the actual result for a couple of weeks.

Also fix some typos in the readme while updating it about the drill
patterns.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>