Add USB version of the LT1945 base board
It's really exactly the same as the DC Jack version, just with the USB-C
connector.

Oh, and I fixed up the edge cuts - for some reason I had done the bottom
cut one mm off the intended symmetry in the Y direction.  So I fix it up
for the boards as I notice.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>