Remove 'Base-' and 'USB-' naming prefix from kicad files
It made sense when there were one or two base boards and it was all
mixed in the top-level directory to separate the base boards from the
effects boards.

But now we have separate directories for (different kinds of) base
boards, so the 'Base-' and 'USB-' prefixes are just redundant useless
noise. Remove it all.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>