Do more edge cut fixups and update for new naming
This fixes all (?) the base boards to have the symmetric outline rather
than missing one mm at the bottom.

In the process it also ends up updating the file-name information inside
the kicad files that came about due to the renames.

Off to manufacturing to see if any of them work.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>