Joiner: make the whole thing a bit saner
Is it _correct_? Who knows.  But now it uses a couple of fairly
straightforward volume controls for the two inputs, and just a regular
op-amp adder to join the two together, instead of trying to balance the
two against each other at the adder.

Maybe this is worth manufacturing now?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>