Add SOT-23-THIN-14 footprint
This exists in the standard KiCad footprint libraries for the 16-pin
version, but not the 14-pin one.  Strange.

The 14-pin version is for a TI SN74HCS74DYY - dual D-type flip-flop with
Schmitt trigger inputs.  I'm using the normal 74*74 symbol (same
pinout), just with the new footprint that KiCad doesn't seem to have, or
I just couldn't find.

It's auto-generated using the footprint wizard thing.  Which doesn't
generate the nice rounded rectangles for the pads for me for some
reason, but whatever.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>