Add a USB-C base board with latching circuit for a momentary stomp switch
This uses a analog switch/mux, so it's not "true bypass".  And I really
don't think I should use the third analog channel for the LED switching,
but it should work.

I could either drive it directly from the flip-flop (which has less
drive current, but should be fine), or have an actual extra transistor
to drive the LED, but that third analog channel was unused and it's
mirroring what a real 3PDT switch design does, so ...

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>