Octaver: replace the effect switch with a blending pot
Instead of a hard switch between octaver circuit bypass, use a pot and
another op-amp to blend the original and the rectified signal before the
tone control.

To make it all simpler, this removes the DC bias after blending, so that
the blending inputs are both driven similarly directly from the previous
op-amp.

It shouldn't matter - the blending stage will happily blend in the DC
offset and while it conceptually makes more sense after the
rectification stage rather than the blending stage, the end result is
the same.

Let's hope the end result is "it works". That would be good.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>