Add C_0704 footprint for Murata GRMJN series caps
I was looking for reasonable 0.1uF C0G caps on Mouser, because they end
up being used both for the power rails (it's a typical suggested size
for the op-amp bypass caps) and for the signal side.  I don't want to
deal with two different 0.1uF caps for the two uses, thus the "look
around for reasonable cap for both".

The Murata GRMJN65C1H104JE01J ended up looking good, and isn't the
enormous 1206 size that I've been using.

It's an unusual size (which is part of why I hadn't seen it before).
The "JN" part of the code name means that it's a 0704 footprint (1810
Metric: 1.8mm by 1.0mm).

The 0805 footprint looks like it would work just fine, but here's an
attempt at something better.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>