Allow trimming gain on base board input buffer
Instead of just having a unity gain buffer, let's add a trim pot that
allows for higher input gain when that makes sense for the effect board.

For example, the very traditional Big Muff fuzz pedal seems to have an
input stage with roughly a voltage gain of 7x (+17dB).

Random pick: 100k trim pot with 10k resistor gives a gain range of 1-11.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>