Update Muffish board for random kicad file changes
The kicad file format isn't very stable, and it does things like
sometimes save rotation angles as 360°, and then when you load it and
re-save it, it turns it into the saner 0° angle.  Similar issues with
fill positioning etc.

But maybe I did move something around in there, so commit the noise.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>