Random guitar pedal project: The Next Generation
This is a reset of the guitar pedal project based on what building the
first generation of 1590A pedals taught me.

The original is still available in the 'legacy' branch, but this has new
board connector locations, and new drill patterns (darn - new enclosures
needed).

The likelihood that this works may be higher, and it's going to fit
better in the inclosure, but it's still unlikely to be very _useful_ as
a guitar pedal.

There's also a new design goal: now that I know the enclosure fits, and
I have picked a base board design, I want to actually debug the effect
boards.

In particular, the original rectifier board ended up not working very
well, and was hard to debug.  But I really don't like breadboarding, and
don't have the through-hole versions of my SMD components anyway.

So the plan is to a hybrid version, which avoids breadboarding for the
basics, but can connect to a "breadboard" for testing.  But right now
it's more of a "basic board with test points" thing.

And in the meantime, here's the VGND base that worked well, along with
the legally mandated boost effect, and a new trial at the rectifier.
Because do I learn from my mistakes? No.  I just try again.

Committing this in this form because I don't have time to really work on
it, but I'm sending it off for manufacturing in the hope that by the
time I get boards back the kernel merge window will be over and I'll
have time to play with it.  Need to think about the breadboarding more,
because this really isn't it, even if there are a few more test-points.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>