Oops, fix the 4.5V ref voltage - remove the coupling capacitors
The bypass caps came from the "real ground" circuits that had dual
voltage rails, but they are very wrong for the 4.5V ground generated by
the TLE2426.

There's even a big comment saying that it's a voltage reference, not a
power supply, but I didn't think it through.  Having the caps to the
real voltage rails just makes the 2426 oscillate, so don't do it.

The good news? The boards work, just make sure to not populate the caps.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>