Stop trying to pretend that the 4.5V reference voltage is ground
Yes, it was a real ground in the earlier dual rail supply designs, but
in Gen2 that's no longer the case, and the pretense caused issues with
me adding bypass capacitors to that reference voltage rail.

So this just renames it as 'GNDREF' everywhere.  Maybe I should have
called it 'VREF'?

The real ground is still called '-V' in the effects, I'll change that
later too, but let's do this in individual steps.

This also removes the remaining bypass caps to GNDREF, and makes the
ground pours be actual ground.

I also need to come up with something sane for the Rectifier board
potentiometers.

Right now they are bogus random stuff where I am using the pots as
trim-pots for testing, but they don't actually make *sense* (ie the
"Volume" pot isn't actually sane, it ends up being another voltage
divider with the resistors in the low-pass filter trim pot etc).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>