Remove breadboard base and rectifier boards
It turns out that it worked fine, but I'm trying to make more of a debug
infrastructure than that initial "breadboard" setup was (which never
ended up being connected to any breadboard).

In particular, I want to first add some connectors for better interfaces
to the signal generator and oscilloscope, and then the regular VGND base
can act as the debug base, rather than have a separate board for it.

And instead of this kind of "effect board debug" thing, I want to
actually then connect the base board to a breadboard.  Exact details
fuzzy, but I did like having these test boards to actually check the
signal, but it wasn't flexible enough to really do anything else.

And I think I can test the signals almost as well using just a simple
connector board.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>