Add connector debug board
As mentioned, this was the main advantage of the breadboard test boards:
just having much better connections to the signal generator and
oscilloscope.

So the old breadboards are gone, and this does *just* that trivial
connector board thing instead.

It's literally just a "audio jack connected to a BNC connector", except
it's two of them side by side with switch to connect them - or not.

IOW, you can either use it as two independent BNC<->Audio connectors
(for the input and output sides of a pedal circuit under test), or
joined together as one single "measure the signal in the middle of the
audio path" board.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>