Fix up a couple of the symbols in 'Mylib'
Add footprint and value to the OPA1655, and proper value to the random
P-channel MOSFET I'm using (PMV250EPEA).

Also add the BAT41ZFILM diode I'm using.  It's the SMD version of the
BAT41 that KiCad already has and that I was using, but the KiCad
standard library version is the THT one, and if I do a "update from
library", it will replace the footprint with the THT footprint version
because that's the only one KiCad knows about.

So now I made a duplicate SMD diode symbol just for this.  There's
probably some proper fix for this.  Maybe I should just have used a
generic Schottky diode symbol that KiCad doesn't then force the wrong
footprint for?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>