Trivial updates to the VGND board
This is mainly just some added experimental noise suppression, with an
optional ferrite bead on the power ground connection too.

None of this should matter - or necessarily even be a good idea - for a
good pedal power supply with clean power, but this is mostly just a
reaction to a not-so-good no-name power supply that was very noisy
indeed.  It's just been pending while I was busy with the kernel merge
window.

I threw that power supply out in disgust, which was a mistake.  I should
have kept it for testing this.

But even if the ferrite beads don't end up working or making any sense,
it's easy enough to just bridge them.  It's hard to try to add them to a
board that doesn't have the space for them.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>