Update boost pedal for misc minor changes
This turns the Bass pot around.  It used to increase Bass when turning
the pot up, but I feel like it makes more sense to turn the Baxandall
tone circuit logic to be about "widening" or "narrowing" the frequency
range.

So now the bass pot turns the bass down when turned clockwise, and the
treble pot turns the treble down when turned counterclockwise
("narrowing" the band), while widening the band obviously works the
other way.

Does that maker sense? Maybe.

Also change the board to use the new BAT41ZFILM symbol.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>