Add Tremolo board for testing
This is about as simple as I could make it using the same components
I've used for other things.

So it uses an opamp as a very simple LFO, which then drives a transistor
that ends up shorting the signal to ground.  Add another transistor for
a blinky LED to give visual indication of the speed.

Does this work? It simulated ok.  But even if it works, this is not some
kind of nice, smooth, pleasant tremolo.  This is a harsh and hard square
wave thing.  Which made the LFO easy, and means that I can largely
ignore any worries about transistor biasing or whatever, because the
driving signal isn't exactly subtle.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>