Add back a LT1054-based board with +-9V rails
The rail splitter has worked really well for the boost effect, and the
power supply filtering does seem to have made the power rails really
quite clean too.  Looking at them with the scope it all looks good.

But for some reason the rectifier board really messes with the rail
splitter.  I get oscillations in the virtual ground at ~2.5MHz with the
rectifier board that I don't see with the boost board at all.  I thought
I had that fixed with putting the power capacitors just between VIN and
real ground, and the boost vboard really looked fine - but no.

I may be doing something horribly wrong and it's something else
entirely, but I'm starting to suspect that I'm hitting the documented
unstable region in the output current vs load capacitance range of the
TLE2426 (fig 17 in the data sheet).

Reading the stability range graph I should be able to fix that by adding
_more_ capacitance, what I used to have also caused issues and should
actually have been outside that unstable range.  So there is clearly
something I don't understand going on.

Which is all good - the point of the project was to learn.  But just to
verify that it really is about the TLE2426, I am doing a simplified
version with just the TL1054 instead.  Because that way signal ground
will be the real global ground, and if I see oscillations I know
something else is going on.

.. and if that turns out to be the issue, I guess that also solves the
headphone amplifier thing, and I can go back to using the BUF634 on the
output without an output capacitor.

And yes, if it really is that I'm hitting the TLE2426 instability
region, clearly I would have been better off just using a plain resistor
divider thing. I thought I was doing something better.

The oscillation frequency is clearly way beyond the range of any audio
signals, so it might not "matter" in the end, but since I certainly
cannot rely on my ears to tell if the effects are doing the right thing,
I do need the whole "signal generator and oscilloscope" thing to be able
to make any real judgment.

The oscillation just makes that measurement model not work at all, so
it's a deal-breaker whether it's audible or not.

This converts just the problematic rectifier board to the new world
order.  I'll do the others for more coverage for testing too: they may
work as-is (because they still see "GNDRF / +V / -V" on the same pins),
but with a real ground I should put power caps sifferently (and the
voltages for the LED on the tremolo board is different, even if it
should still work).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>