Add back the BUF634A to the audio output
With the dual rails, we don't need to have the output capacitor to
rebase the signal levels, and can drive it directly.  Which allows going
back to the headphone amplifier setup using the BUF634A.

Is this a good idea? I dunno.  Having a separate headphone amplifier
makes more sense, I guess, but I liked just being able to connect the
headphones to test things out.

And it's not like these boards make sense for anybody else anyway, so...

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>