Tremolo: hopefully fix stupid oversights
So it turns out that the popping noise in the first version of the
tremolo circuit was something that I *should* have seen in simulation,
but didn't.  I blame incompetence, but the excuse is that the spike in
output was hidden by the control voltage signal that I also simulated,
and that overlapped it.

But that gave me a clue to how to fix it - don't use the very abrupt
control voltage from the opamp output, instead use the voltage from the
capacitor charge.  Same LFO, just different point in the circuit.  And
the capacitor voltage isn't a hard square wave, but looks more like a
triangle wave, and doesn't cause that kind of nasty harsh edge.

Also fix the transistor that drives the LED.  It was the wrong way
around, and the body diode of the mosfet would just keep the LED lit all
the time, just with a slightly lower voltage.  Blush.

Of course, now that I'm thinking about that body diode, that also means
that using a MOSFET for the signal switching is probably a bad idea to
begin with.  That very same body diode will also clip the signal if it's
larger than the forward voltage drop.

It tested fine in simulation, and looked fine on the scope, though.  And
sounded fine too, apart from the popping noise.  So the forward voltage
drop is clearly high enough to not be a big deal in practice.

PMV250EPEA datasheet says V_SD is typically -0.8V (max -1.2V).  Which
explains the lack of clipping in practice.

Put two MOSFETs in series in opposite directions with the same gate? Or
use some 4-terminal thing? The reason I'm using that P-MOSFET is that I
already had the device...  Ponder, ponder.

In the meantime, let's fix the actual visible / audible problems, and
not worry about the clipping yet.

Or maybe tout it all as a feature: it's a tremolo pedal with a built-in
metronome from the popping, *and* it does a cool transistor-based
distortion.

It's a floor wax and a dessert topping!

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>