Add simulation details for the components I use regularly
This adds simulation details for the MMBD1203 dual AKC diode, the
PMV250EPEA P-channel MOSFET, and the OPA165x opamps.

NOTE! The actual SPICE files for the devices are not included, they have
to be fetched from the TI website.  Annoying, but there you are.  But
this adds the scaffolding so that _all_ you need to do is fetch them.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>