Make the rectifier board simulate properly in KiCad
This also adds the 5pF capacitor to the feedback of the Baxandall
circuit: that calms it down and stops it oscillating in simulation (and
the simulation seems to actually match reality, unlike the circuitlabs
simulation I did before).

Thanks to Eric Kuzmenko (@gralco) for pushing me to simulate inside of
KiCad with proper models instead of my lazy "this general vicinity" web
simulation.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>