Clean up some more of the simulation description details
This uses relative pathnames for the simulation data, and adds another
couple of "use sane pin names and numers" indirection files that paper
over some of the SPICE file oddities of the manufacturer data files.

There must be some standard for spice files that makes the pin names be
completely arbitrary, but whatever that standard is, KiCad doesn't know
about it, so the simulation pin configuration needs to be done per
symbol.

And to make that simulation pin configuration sane, this adds wrappers
with sane pin ordering and names for the PMV250EPEA and the OPA1655.

The wrappers also _might_ also maybe make it a bit easier to put the
vendor SPICE lib files in different places by just having all of that in
that one wrapper file per symbol, instead of having it spread out in
random KiCad places.

Use ${KIPRJMOD} to make things relative to the *.kicad_pro project file
as suggested by Eric Kuzmenko.

There may be better ways to do this, but this should at least be an
improvement over having hardcoded paths to

   /home/torvalds/KiCad/1950A/...

in the project.

I'm still hitting the occasional "Time step too small" error, so it's
not like all this simulation stuff is solved, but on the whole it's
certainly better than the separate web-based sim approximations I used
to do, and seems way more accurate.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>