Make the tremolo board simulate under KiCad too
The simulation shows that it's working now (modulo the usual "Time step
too small" issues that hit more or less randomly), but also that I'm
clearly missing something.

In paticular, the capacitor in the LFO charges down to -4V (with a +-4.5
supply, in that VGND model), which is what I would expect from the
100k:10k voltage divider in the LFO.

But it only charges *up* to +2.2V before the LFO switches.  Which wasn't
what I was expecting, and it also results in a lower duty cycle (and
higher frequency) of the tremolo effect.

I assume the simulation is right, and the "what I expected" is just
entirely wrong.  That said, the circuitlab.com simulation that I used
previously *did* actually behave like I expected, so it's not like my
expectations were entirely crazy.

One difference is that the KiCad simulation now uses the actual
manufacturer SPICE models for the particular components I'm using.

Maybe the PMV250EPEA acts sufficiently differently from some other
P-MOSFET? Or maybe it's the exact details of the OPA1655 that matter.

I'll have to investigate, but at least this gives me a working
simulation to look at.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>