Simulate the diode selection switch in the boost circuit
You'd think that you would simulate it as a "switch" device.  LOL no.
That would be too obvious.

I'm telling KiCad that it's a 10M potentiometer.  The "off" position is
with the "wiper" in the middle (so 5MOhm to both sides - off enough),
while wiper at 0/1 picks one of the diode clipping options.

Also update the V(IN) to be a 200mV amplitude oscillation (400mV PtP).
Together with the boost pot at 50% that shows clear clipping.

Both diode paths look very similar in simulation, although in real life
the BAT41Z sounds clearly different from the MMBD1203.  The BAT41Z sim
is just the default diode setting, though, so it might be that - or it
might be the particular random boost setting too.

Anyway, this makes it simulate something fairly close to reality, and
while I'm not loving the KiCad simulation oddities (a pot as a SPDT
switch? Really?) and the horrid error messages, I do appreciate being
able to simulate the actual real circuit.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>