Remove the LT1054 board again
The reason for this board existing was that I was wondering whether the
LT2426 rail splitter was oscillating, and wanted to have a guaranteed
solid ground.

But as part of doing all the simulation work due to issue #4 ("Tremolo
doubling as a metronome") it became clear that the oscillation was
actually from the Baxandall circuit.

Now, that on its own might not doom the LT1054 board, and the ability to
add back the headphone buffer _was_ a nice side effect of not doing the
VGND approach.

But it turns out I messed something up when I resurrected and simplified
the LT1054 board, and since the oscillation issue was solved, it's
simply not worth debugging.

The LT1054 board used to work fine back in the original form when I also
made it do the voltage doubling and generated roughly +-15V rails.  But
when I resurrected it for testing, I wanted to make it simpler, and
removed the voltage doubling and just made it do a plain voltage
inversion.

And in that process I clearly missed something, and the test boards I
just got back did not work at all.  The LT1054 chip just got very hot.

Did I connect the transistor that was supposed to protect the -VOUT pin
from ever going positive the wrong way around? Is it the wrong
transistor to use here at all? Very likely something like that.

The original - working - LT1054 board had a simple diode there (well,
more than one actually, because of the voltage doubler circuit and
because the voltage drop over the diodes was actually a good thing in
that situation).

This new and broken circuit was supposed to just do what the LT1054
datasheet suggests in Fig 30, but I note that I messed that up and put
the output capacitor on the wrong side of the transistor.

And the transistor I'm using is a low-current one, so regardless of
whether I put it the wrong way around or not, the choice of transistor
was a mistake too.  It just happens to be the only regular NPN I had
around.

I'm not even entirely sure why I have that random BC846 in my component
library, but I have a sneaking suspicion that the reason is just that
it's the cheapest NPN BJT transistor on mouser and I had been looking at
transistors for signal amplification, not for power supplies.

It's not the first time I use a particular component just because I
happen to have it around, rather than because it's the right thing to
use.  See the use of the P-MOSFET in the tremolo circuit.

Anyway, there could be any number of technical reasons for the failure,
but in the end it comes down to one simple but embarrassing root cause:
my incompetence.

Did I learn anything from this? Not really.  Instead I'm just deleting
this board and am going to pretend it never happened.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>