Update all the PCB's with the recent changes from simulation
Time to get another set of test boards manufactured, this time perhaps
with the high-frequency oscillation fixed for good thanks to much
improved simulations.

And maybe the tremolo works with blinking LEDs too.  I added another
one, because I'm using the same enclosure for all of these, so I need to
put *something* in the holes that are drilled...

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>