Add note about capacitor sizing for bass guitar version of Baxandall
This should center the adjustment frequencies a much lower frequency
range that is more appropriate for bass guitar.

I think I'll have to make one of the capacitors a 1206 to make it
actually available, because while soldering a 1206 on top of the 0805
footprint works, it's just too fiddly.

Possibly just make it a DNP footprint in parallel. Or even a switch.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>