Add spice simulation file for the MTB2222ADW1
It's a dual NPN transistor package, and while I'm sure there's some way
to describe that in kicad natively, I haven't found it.

So I'm doing the same thing as for the OPA1656: creating a spice file
that refers to the actual manufacturer spice file (not committed, you
need to fetch it separately), and describes how the pins on the chip
connect to the two transistors it contains.

Because the manufacturer SPICE lib file only describes one transistor,
and if I just point to that spice file, I can't simulate the second one
without this indirection.

Why this new sim file? Because I'm replacing the op-amp in the tremolo
circuit with two transistors.  And I do not want to solder those things
individually (the same way I prefer that dual-diode MMBD1203 package, I
much prefer just pairing up the transistors in one single package too).

Yes, there are quad packages, and I could do that too, but the dual
transistor package seems fairly reasonable.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>