Update the tremolo board
This basically re-implements it almost from scratch.  Instead of an
op-amp for the LFO, this uses a dual transistor oscillator instead,
because when I was looking into how to adjust the duty cycle of the
square wave, it became clear that this is the much simpler circuit for
that.

Using an op-amp was always a bit of an overkill.  A simple
two-transistor package in a SOT-363 is tiny and costs pennies.  But I
started out building this with things I had in my parts library, and
that was the P-MOSFET I use for reverse current protection, and the op
amps.

Just replace both of those uses with two of these dual NPN BJTs.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>