Updated Tremolo board, version N+1
I'm not sure this is a particularly good tremolo, but it's been an
absolutely stellar board for learning about transistors.

It started out using an op-amp for the oscillator, and the P-MOSFET that
I used for reverse voltage protection as the attenuator.  And that
worked, with some popping noise ("Tremolo doubling as a metronome" #4).

Which was fixed with a nicer driving curve, and the popping went from
"metronome" to "there's still some 'fizz' sound modulated by the
oscillator frequency.  But I also wanted to make sure I could change the
timing more conveniently and have different duty cycles, so I ignored
that and went back to the drawing board.

Then the op-amp circuit was replaced with just a dual transistor
bistable oscillator instead using dual BJT transistors, because that
made the timing much simpler and more flexible.  And in the process, I
did the signal attenuation and the LED driving with another dual BJT,
because why not?

And then the third stage answers the "Why not?" question.  Because it
turns out the BJT as the transistor pulling down to ground didn't work
so well and that was likely also the cause of the 'fizz' with the MOSFET version.

So now I'm using a JFET for that, a separate BJT for the LED, and the
dual BJT still does the oscillator part.

I've got a couple of different JFET's on order: the MMBF4393LT1G was
suggested by Eric Kuzmenko (@gralco), and looks good.  And after looking
at way too many other ones tryin gto figure out the differences, I'm
also going to try the MMBF5103.  Same D/S/G ordering in a SOT-23, so
they are interchangeable.

The MMBF5103 simulates slightly better and hopefully this also fixes the
fizz issue for good.  But let's see if that actually shows up in
reality.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>