Add a USB-C pedal power board
I've given up on actually using USB-C for power in the pedals
themselves, but I end up having the components for it left over from my
trials, and I still prefer having USB-C on the bench over some random
guitar power wall-wart for testing.

So I made a pedal power board using the component library I have around.

This thing is not even _attempting_ to keep ports isolated: normally I
only have the one board under test anyway, but since I occasionally test
two in series and the LDO is supposed to be able to do up to 1A at 9V, I
put several barrel jacks on this thing.

It might be too noisy to be useful, but hey, it's another small
soldering project, and if it works out ok it migth just remove a
wallwart.

This uses the HUSB237 USB-C PD sink controller to get 12V out of a USB-C
PD power supply (at up to 2.5A), and then the TPS73801 LDO to turn that
into 9V (at up to 1A).

I got the DFN2x2-6L version, which has those data lines that I don't
even use (because I use the 6-pin USB-C connector that is power-only)
and is a slight pain to solder, but it's doable.  I couldn't find the
SOT-23-5 version of the HUSB237 anywhere, but that really seems
basically the perfect chip for simple USB-C PD sink applications.

  https://www.lcsc.com/datasheet/lcsc_datasheet_2410121803_Hynetek-HUSB237-AA001-DN06R_C22373734.pdf

It's ridiculous how simple the HUSB237 is to use for these kinds of "use
USB C PD as a barrel jack replacement" kind of applications, when the
big US companies all make totally over-designed garbage that needs I2C
to drive the choice etc.

You simply connect a resistor or a capacitor between VSET and GND to say
what you want the HUSB to negotiate (*), and you're done (or leave it
floating for 12V like here).  No switching transistors needed for the
power side, because all you get is VBUS.  Dead simple.

 (*)	 5V: 0R
	 9V: 75k
	12V: leave floating
	15V: 4.7nF
	20V: 20nF

I'm sure the TI's of the world think the HUSB237 is a horribly hacky
thing, and don't want to compete in the low-end market.  But honestly,
the only thing I would have wanted from the HUSB237 that it didn't have
was a "PD negotiation confirmation" signal.  I'd love to have a LED that
shows "yes, you got the 12V you asked for".

Now it does have a LED that will probably be too dim if it didn't get
12V, so the board does have a status signal.  I guess that was the
thinking: the voltage you get *is* the confirmation.

Sadly, this cool little USB C PD sink chip isn't available from Mouser
and the like.  I got it from LCSC.

Anyway, I should probably have made this board be switchable to
optionally do 12V output instead (request 15V from USB-C, switch
resistors for the TPS73801 around a bit) just as a test-case for my
pedals.

If this ends up not being too noisy to use, I'll probably update it.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>