Version N+1 of the rectifier board
I have enjoyed the design side of the precision rectifier circuit.  The
concept of the circuit - rater than the resulting distortion - was
admittedly why I started it in the first place.

The idea was that the rectifier could be used for other things than as
the actual signal output (ie using it as the peak detector in a
compression pedal or whatever).  But using it as the distortion itself
was kind of the first step.

But I have to admit to really not liking the *sound* of the resulting
octaver fuzz.  I think it's too much fizz, and the alleged octave-up
side is entirely drowned in all the high-order harmonics.

So I have a cunning plan: replace the Baxandall tone control with a
rather aggressive second-order low-pass filter.  That should counter-
act the harsh signal spikes when the polarity of the signal changes, and
for some frequencies will (at least in simulation with a pure sine-wave)
result in a pretty clean octave-up effect.

Will this be a better sound? Who knows.  But it's certainly a different
one.  And the Baxandall tone circuit was already done - and worked
fairly well - in the regular diode-clipping overdrive pedal.

This also tunes the diode choice for the rectification a bit, simply
because I started looking at different diodes in simulation.  A
low-V_{f} Schottky diode simulates better in that critical polarity
change region, but who knows how accurate the spice models here are.

Anyway, for the common cheap SOT-23-3 diodes in series (so that I can
just try different footprints) the BAT54S simulated best out of the ones
I tried.  The Rohm RB706WM-40 simulated even better, but doesn't exist
in SOT-23-3.

Put another way: this is another step in the "Linus learns about random
analog circuits".  This time the new circuit being a low-pass Sallen-Key
filter to hopefully make the octaver fuzz pedal sound better.  Or at
least sound different.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>