Boost pedal: try the BAT54S here too
The boost pedal used to have discrete BAT41 Schottky diodes, but since I
got the SOT-23 BAT54S fro the rectifier board, let's see how that works.

The BAT54S might actually be *too* aggressive, since it hass a very low
V_{F} indeed.  But hey, people like and use germanium diodes in clipping
circuits, so think of it as a modern version of that.

Of course, in guitar pedals, it's heretical to not use germanium diodes
that you have hand-picked, along with film capacitors.  People are
convinced it's a special softer kind of clipping, and yes, I'm sure the
IV curve looks different.

But hey, this is all inside the op-amp feedback loop anyway, so it's a
soft clipping regardless.

More importantly, using the BAT52S means that there's one less annoying
polarized component to worry about. Two, actually.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>