Add a LDO version of the VGND board
I call this the "digital" board, because that's the impetus for this:
I'm thinking of doing a delay pedal using the pt2399, and that wants a
5V power rail with a 2.5V reference.

I could obviously do all that just on the effects board with simple 5V
LDO, another 2.5V reference voltage, all the signal re-biasing etc.  But
that's *literally* what the base board exists for, so...

While a very simple fixed 5V LDO would work fine (or just the LM7805:
old-school) - I already have TPS73801 in my parts library, and it does
the reverse polarity protection too.  So while it adds a chip, it then
removes the need for all the other crud.

[ Side note: I really wish I had noticed the TL1963A before getting the
  TPS73801. Not because of the higher max current, but because of how
  annoying I find the EN pin on the TPS73801.

  In typical use, you tie the EN pin high (the TL1963A calls it a
  shutdown pin, active low, but same deal). And whoever specified the
  pin-out of the TPS73801 decided that the logical way to do that is
  obviously to put the two pins ON OPPOSITE ENDS OF THE CHIP!

  The TL1963A has a much simpler pinout with IN and SHDN right next
  to each other.

  Yeah, it's just one extra trace that I put on the other side of the
  board because I want to keep the power traces as solid and simple as
  possible. I don't know why I find it so annoying, but I do. ]

If I ever want to try some modern 3.3V thing, I'd be screwed - I can't
really go lower than 5V without having to re-do all the effects boards
and picking another op-amp since the OPA165x wants at least a 4.5V
supply.  But a 5V rail is fine and should work as-is with all the
existing effects boards.

And I could go the other way and make a 8V board by just changing the
resistor divider, and maybe the TPS73801 would then deal better with
noisy power supplies? Who knows?

So this *might* become the standard board and replace the VGND one
entirely, but for now it's a separate one just for testing, and as prep
while I ponder the PT2399.

I've simplified some of the other components on this board while at it
and tried to make this pretty minimal.  And I decided I really want an
C0G cap on the whole signal chain - the output re-biasing in particular.
And that means 0.47uF is the biggest I can reasonably do.

I'm not sure I need even that, and the same 0.1uF that is on the input
should work, but I think aiming for fairly low output impedance is the
name of the game.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>