Add a custom 9mm pot footprint
I've been using the existing KiCad footprint for the 9mm alpha pots I
use, but they have a couple of issues that I found annoying:

 - the footprint is centered around pin #1, which makes it hard to place
   correctly and particularly to rotate. Sure, pin1 kind of makes sense
   on the board, but for these things the _primary_ positioning is the
   position of the shaft, not the through-hole of one of the pins

 - the footprint doesn't have nets for the two mounting holes, so if you
   try to connect them to ground, that will make DRC very unhappy

 - the standard footprint has the full "proper" courtyard, which makes
   a lot of sense, but is annoying when I would want to place low
   profile SMT components underneath it. Again, I can obviously just do
   that anyway, but I really don't like the resulting DRC errors.

So this makes a new custom footprint for the pots I use with these
issues all fixed.  Use the pot with the mounting pin, and connect that
pin to ground or shield or whatever.

And the courtyard is very limited indeed now, so it's very permissive.
Obviously that can cause the reverse kinds of problems where DRC doesn't
warn about insane component placement, but hopefully that will be so
visually obvious that it's not a problem in practice.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>